+++ src/Network/Network/ChangeLog.md
@@ -21,7 +21,7 @@
 ## Upcoming Release
 * Added support for equal-cost multi-path (ECMP) routing in route tables
     - Added the `VirtualApplianceEcmp` next hop type to `New-AzRouteConfig`, `Add-AzRouteConfig`, and `Set-AzRouteConfig`
-    - Added the `-NextHopIpAddresses` parameter to specify the list of next hop IP addresses for ECMP routes (minimum of 2 addresses)
+    - Extended the `-NextHopIpAddress` parameter to accept a list of next hop IP addresses (2 to 64) for ECMP routes when the next hop type is `VirtualApplianceEcmp`
     - Added the `NextHop` property to the route output returned by `Get-AzRouteTable`
 * Added property 'Nat64' to NatGateway and support for it in the following cmdlets:
     - `New-AzNatGateway`